Big commit. Restructring things to provide a channel of stashes.
Don't like how large this one is, but all these changes are so closely
coupled it'd be hard to do a bunch of coherent small patches.

StashStream exposes Start() which fires a goroutine that loops,
requesting each new page of the stash tab data, then feeds the items one
by one into a channel for processing down the line.

Non-working.